+++ src/request.rs
@@ -221,17 +221,16 @@ impl<'r> core::iter::FusedIterator for PathSegments<'r> {}
 
 /// Represents an HTTP request.
 #[derive(Debug, Clone, Copy)]
-pub struct Request<'r> {
+pub struct RequestParts<'r> {
     method: &'r str,
     path: Path<'r>,
     query: Option<UrlEncodedString<'r>>,
     fragments: Option<UrlEncodedString<'r>>,
     http_version: &'r str,
     headers: Headers<'r>,
-    body: &'r [u8],
 }
 
-impl<'r> Request<'r> {
+impl<'r> RequestParts<'r> {
     pub fn method(&self) -> &'r str {
         self.method
     }
@@ -255,10 +254,187 @@ impl<'r> Request<'r> {
     pub fn headers(&self) -> Headers<'r> {
         self.headers
     }
+}
 
-    pub fn body(&self) -> &'r [u8] {
-        self.body
+pub struct RequestBodyReader<'r, R: Read> {
+    content_length: usize,
+    reader: &'r mut R,
+    current_data: &'r [u8],
+    read_position: &'r mut usize,
+}
+
+impl<'r, R: Read> crate::io::ErrorType for RequestBodyReader<'r, R> {
+    type Error = R::Error;
+}
+
+impl<'r, R: Read> RequestBodyReader<'r, R> {
+    pub fn content_length(&self) -> usize {
+        self.content_length
     }
+}
+
+impl<'r, R: Read> Read for RequestBodyReader<'r, R> {
+    async fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
+        let read_size = if self.current_data.is_empty() {
+            let max_read_size = buf.len().min(self.content_length - *self.read_position);
+
+            if max_read_size == 0 {
+                0
+            } else {
+                self.reader.read(&mut buf[..max_read_size]).await?
+            }
+        } else {
+            let read_size = self.current_data.len().min(buf.len());
+
+            let (current_data, remaining_data) = self.current_data.split_at(read_size);
+
+            buf[..read_size].copy_from_slice(current_data);
+            self.current_data = remaining_data;
+
+            read_size
+        };
+
+        *self.read_position += read_size;
+
+        Ok(read_size)
+    }
+}
+
+#[derive(Debug)]
+pub enum ReadAllBodyError<E> {
+    BufferIsTooSmall,
+    UnexpectedEof,
+    IO(E),
+}
+
+#[derive(Debug)]
+pub struct RequestBody<'r, R: Read> {
+    content_length: usize,
+    reader: &'r mut R,
+    buffer: &'r mut [u8],
+    read_position: &'r mut usize,
+    buffer_usage: &'r mut usize,
+}
+
+impl<'r, R: Read> RequestBody<'r, R> {
+    pub fn content_length(&self) -> usize {
+        self.content_length
+    }
+
+    pub async fn read_all(self) -> Result<&'r mut [u8], ReadAllBodyError<R::Error>> {
+        let buffer = self
+            .buffer
+            .get_mut(..self.content_length)
+            .ok_or(ReadAllBodyError::BufferIsTooSmall)?;
+
+        if let Some(remaining_body_to_read) = buffer.get_mut(*self.buffer_usage..) {
+            self.reader
+                .read_exact(remaining_body_to_read)
+                .await
+                .map_err(|err| match err {
+                    embedded_io_async::ReadExactError::UnexpectedEof => {
+                        ReadAllBodyError::UnexpectedEof
+                    }
+                    embedded_io_async::ReadExactError::Other(err) => ReadAllBodyError::IO(err),
+                })?;
+
+            *self.buffer_usage = self.content_length;
+        }
+
+        *self.read_position = self.content_length;
+
+        Ok(buffer)
+    }
+
+    pub fn reader(self) -> RequestBodyReader<'r, R> {
+        RequestBodyReader {
+            content_length: self.content_length,
+            reader: self.reader,
+            current_data: &self.buffer[..(self.content_length.min(*self.buffer_usage))],
+            read_position: self.read_position,
+        }
+    }
+}
+
+pub struct RequestBodyConnection<'r, R: Read> {
+    content_length: usize,
+    reader: &'r mut R,
+    read_position: usize,
+    buffer: &'r mut [u8],
+    buffer_usage: usize,
+    has_been_upgraded: &'r mut bool,
+}
+
+impl<'r, R: Read> RequestBodyConnection<'r, R> {
+    pub fn content_length(&self) -> usize {
+        self.content_length
+    }
+
+    pub fn body(&mut self) -> RequestBody<'_, R> {
+        RequestBody {
+            content_length: self.content_length,
+            reader: self.reader,
+            read_position: &mut self.read_position,
+            buffer: self.buffer,
+            buffer_usage: &mut self.buffer_usage,
+        }
+    }
+
+    pub async fn finalize(
+        self,
+    ) -> Result<crate::response::Connection<'r, impl Read<Error = R::Error> + 'r>, R::Error> {
+        // If the entire body is already in the buffer
+        if self.content_length <= self.buffer_usage {
+            return Ok(crate::response::Connection {
+                reader: crate::response::BufferedReader {
+                    reader: self.reader,
+                    buffer: self.buffer,
+                    read_position: self.content_length,
+                    buffer_usage: self.buffer_usage,
+                },
+                has_been_upgraded: self.has_been_upgraded,
+            });
+        }
+
+        // Data after the body has not yet been read, the entire buffer can be used to read the rest of the body
+
+        // Skip the section that has already been read into the buffer
+        let mut read_position = self.read_position.max(self.buffer_usage);
+
+        while let Some(data_remaining) = self
+            .content_length
+            .checked_sub(read_position)
+            .and_then(core::num::NonZeroUsize::new)
+        {
+            let read_buffer_size = data_remaining.get().min(self.buffer.len());
+
+            let read_size = self
+                .reader
+                .read(&mut self.buffer[..read_buffer_size])
+                .await?;
+
+            if read_size == 0 {
+                break;
+            }
+
+            read_position += read_size;
+        }
+
+        Ok(crate::response::Connection {
+            reader: crate::response::BufferedReader {
+                reader: self.reader,
+                buffer: self.buffer,
+                read_position: 0,
+                buffer_usage: 0,
+            },
+            has_been_upgraded: self.has_been_upgraded,
+        })
+    }
+}
+
+pub struct Request<'r, R: Read> {
+    pub parts: RequestParts<'r>,
+    pub body: RequestBodyConnection<'r, R>,
 }
 
 #[derive(Debug)]
@@ -268,34 +444,58 @@ pub enum ReadError<E> {
     Other(E),
 }
 
-pub struct Reader<'b, R: Read> {
+pub(crate) struct Reader<'b, R: Read> {
     reader: R,
     read_position: usize,
     buffer: &'b mut [u8],
     buffer_usage: usize,
+    has_been_upgraded: bool,
 }
 
 impl<'b, R: Read> Reader<'b, R> {
-    pub async fn new(
-        mut reader: R,
-        buffer: &'b mut [u8],
-    ) -> Result<Option<Reader<'b, R>>, R::Error> {
-        let buffer_usage = reader.read(buffer).await?;
-
-        Ok((buffer_usage > 0).then_some(Self {
+    pub fn new(reader: R, buffer: &'b mut [u8]) -> Self {
+        Self {
             reader,
             read_position: 0,
             buffer,
-            buffer_usage,
-        }))
+            buffer_usage: 0,
+            has_been_upgraded: false,
+        }
+    }
+
+    fn wind_buffer_to_start(&mut self) {
+        if let Some(used_buffer) = self.buffer.get_mut(..self.buffer_usage) {
+            used_buffer.rotate_left(self.read_position);
+
+            self.buffer_usage -= self.read_position;
+        } else {
+            self.buffer_usage = 0;
+        }
+
+        self.read_position = 0;
+    }
+
+    pub async fn request_is_pending(&mut self) -> Result<bool, R::Error> {
+        if self.has_been_upgraded {
+            Ok(false)
+        } else {
+            self.wind_buffer_to_start();
+
+            if self.buffer_usage > 0 {
+                Ok(true)
+            } else {
+                self.buffer_usage = self.reader.read(self.buffer).await?;
+                Ok(self.buffer_usage > 0)
+            }
+        }
     }
 
     fn used_buffer(&self) -> &[u8] {
         &self.buffer[..self.buffer_usage]
     }
 
-    async fn read_more(&mut self, required_space: usize) -> Result<(), ReadError<R::Error>> {
-        while (self.read_position + required_space) > self.buffer_usage {
+    async fn next_byte(&mut self) -> Result<u8, ReadError<R::Error>> {
+        if self.read_position == self.buffer_usage {
             let read_size = self
                 .reader
                 .read(&mut self.buffer[self.buffer_usage..])
@@ -309,22 +509,10 @@ impl<'b, R: Read> Reader<'b, R> {
             self.buffer_usage += read_size;
         }
 
-        Ok(())
-    }
-
-    async fn next_byte(&mut self) -> Result<u8, ReadError<R::Error>> {
-        self.read_more(1).await?;
         let b = self.used_buffer()[self.read_position];
         self.read_position += 1;
 
         Ok(b)
-    }
-
-    async fn next_slice(&mut self, len: usize) -> Result<&[u8], ReadError<R::Error>> {
-        self.read_more(len).await?;
-        let start = self.read_position;
-        let end = self.read_position + len;
-        Ok(&self.used_buffer()[start..end])
     }
 
     async fn read_line(&mut self) -> Result<Subslice, ReadError<R::Error>> {
@@ -398,36 +586,32 @@ impl<'b, R: Read> Reader<'b, R> {
         }
     }
 
-    pub async fn read(
-        &mut self,
-    ) -> Result<(Request<'_>, super::response::Connection<&mut R>), ReadError<R::Error>> {
+    pub async fn read(&mut self) -> Result<Request<'_, R>, ReadError<R::Error>> {
+        self.wind_buffer_to_start();
+
         let request_line = self.read_request_line().await?;
 
         let request_line = request_line.range();
 
         let headers = self.read_headers().await?;
 
-        let body_length = Headers(headers.as_ref())
+        let content_length = Headers(headers.as_ref())
             .get("content-length")
             .and_then(|value| value.parse::<usize>().ok())
             .unwrap_or(0);
 
         let headers = headers.range;
 
-        let body_start = self.read_position;
+        let parts_length = self.read_position;
 
-        self.next_slice(body_length).await?;
-
-        let body_end = body_start + body_length;
-
-        let used_buffer = &self.buffer[..self.buffer_usage];
+        let (parts_buffer, body_buffer) = self.buffer.split_at_mut(parts_length);
 
         let RequestLine {
             method,
             url,
             http_version,
         } = request_line
-            .index_buffer(used_buffer)
+            .index_buffer(parts_buffer)
             .as_str()
             .map_err(|_| ReadError::BadRequestLine)?;
 
@@ -441,17 +625,31 @@ impl<'b, R: Read> Reader<'b, R> {
                 (Path(UrlEncodedString(path)), Some(UrlEncodedString(query)))
             });
 
-        Ok((
-            Request {
+        let headers = Headers(&parts_buffer[headers]);
+
+        let request = Request {
+            parts: RequestParts {
                 method,
                 path,
                 query,
                 fragments,
                 http_version,
-                headers: Headers(&used_buffer[headers]),
-                body: &used_buffer[body_start..body_end],
+                headers,
             },
-            super::response::Connection(&mut self.reader),
-        ))
+            body: RequestBodyConnection {
+                content_length,
+                reader: &mut self.reader,
+                read_position: 0,
+                buffer: body_buffer,
+                buffer_usage: self.buffer_usage - parts_length,
+                has_been_upgraded: &mut self.has_been_upgraded,
+            },
+        };
+
+        // This will be true once the RequestBodyConnection has been finalized, which happens no matter how the request is handled
+        self.read_position += content_length;
+        self.buffer_usage = self.buffer_usage.max(self.read_position);
+
+        Ok(request)
     }
 }