+++ src/response/fs.rs
@@ -104,30 +104,35 @@ impl File {
 }
 
 impl<State, PathParameters> crate::routing::RequestHandler<State, PathParameters> for File {
-    async fn call_request_handler<W: super::ResponseWriter>(
+    async fn call_request_handler<R: Read, W: super::ResponseWriter<Error = R::Error>>(
         &self,
         _state: &State,
         _path_parameters: PathParameters,
-        request: crate::request::Request<'_>,
+        request: crate::request::Request<'_, R>,
         response_writer: W,
     ) -> Result<ResponseSent, W::Error> {
-        if let Some(if_none_match) = request.headers().get("If-None-Match") {
+        if let Some(if_none_match) = request.parts.headers().get("If-None-Match") {
             if if_none_match
                 .split(',')
                 .map(str::trim)
                 .any(|etag| self.etag == etag)
             {
                 return response_writer
-                    .write_response(super::Response {
-                        status_code: status::NOT_MODIFIED,
-                        headers: self.etag.clone(),
-                        body: super::NoBody,
-                    })
+                    .write_response(
+                        request.body.finalize().await?,
+                        super::Response {
+                            status_code: status::NOT_MODIFIED,
+                            headers: self.etag.clone(),
+                            body: super::NoBody,
+                        },
+                    )
                     .await;
             }
         }
 
-        self.clone().write_to(response_writer).await
+        self.clone()
+            .write_to(request.body.finalize().await?, response_writer)
+            .await
     }
 }
 
@@ -142,7 +147,7 @@ impl super::Content for File {
 
     async fn write_content<R: Read, W: Write>(
         self,
-        _connection: super::Connection<R>,
+        _connection: super::Connection<'_, R>,
         mut writer: W,
     ) -> Result<(), W::Error> {
         writer.write_all(self.body).await
@@ -150,11 +155,14 @@ impl super::Content for File {
 }
 
 impl super::IntoResponse for File {
-    async fn write_to<W: super::ResponseWriter>(
+    async fn write_to<R: Read, W: super::ResponseWriter<Error = R::Error>>(
         self,
+        connection: super::Connection<'_, R>,
         response_writer: W,
     ) -> Result<ResponseSent, W::Error> {
-        response_writer.write_response(self.into_response()).await
+        response_writer
+            .write_response(connection, self.into_response())
+            .await
     }
 }
 
@@ -199,21 +207,21 @@ impl Directory {
 }
 
 impl<State, CurrentPathParameters> PathRouter<State, CurrentPathParameters> for Directory {
-    async fn call_path_router<W: super::ResponseWriter>(
+    async fn call_path_router<R: Read, W: super::ResponseWriter<Error = R::Error>>(
         &self,
         state: &State,
         current_path_parameters: CurrentPathParameters,
         path: Path<'_>,
-        request: crate::request::Request<'_>,
+        request: crate::request::Request<'_, R>,
         response_writer: W,
     ) -> Result<ResponseSent, W::Error> {
-        if !request.method().eq_ignore_ascii_case("get") {
+        if !request.parts.method().eq_ignore_ascii_case("get") {
             return crate::routing::MethodNotAllowed
                 .call_request_handler(state, current_path_parameters, request, response_writer)
                 .await;
         }
 
-        if let Some(file) = self.matching_file(request.path()) {
+        if let Some(file) = self.matching_file(path) {
             file.call_request_handler(state, current_path_parameters, request, response_writer)
                 .await
         } else {